+++ Cargo.toml
@@ -23,6 +23,7 @@ serde = { version = "1.0", default-features = false, features = ["derive"] }
 serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
 embedded-hal = { version = "0.2", features = ["unproven"] }
 dht11 = { version = "0.3.1", optional=true}
+postcard = "1.0.4"
 
 [features]
 default = []
